+++ include/shark/Data/LabelOrder.h
@@ -104,14 +104,14 @@ public:
 		m_labelOrder.clear();
 
 		// and one array that tracks what we already encountered
-		std::vector<int> foundLabels(maxLabel - minLabel + 1, -1);
+		std::vector<unsigned int> foundLabels(maxLabel - minLabel + 1, -1);
 
 		// and insert all labels we encounter
-		size_t currentPosition = 0;
+		unsigned int currentPosition = 0;
 		for(std::size_t i = 0; i < dataset.numberOfElements(); i++)
 		{
 			// is it a new label?
-			int label = dataset.labels().element(i);
+			unsigned int label = dataset.labels().element(i);
 			if(foundLabels[label - minLabel] == -1)
 			{
 				foundLabels[label - minLabel] = currentPosition;
@@ -124,7 +124,7 @@ public:
 		for(std::size_t i = 0; i < dataset.numberOfElements(); i++)
 		{
 			int label = dataset.labels().element(i);
-                        dataset.labels().element(i) = foundLabels[label - minLabel];
+			dataset.labels().element(i) = foundLabels[label - minLabel];
 		}
 	}
